Mood swing flask/requirements.txt
Flask==3.0.2
assemblyai
python-dotenv